Application/requirements.txt
alembic==1.5.7
click==7.1.2
decorator==4.4.2
dnspython==2.1.0
email-validator==1.1.2
Flask==1.1.2
flask-kafka==0.0.3
Flask-Login==0.5.0
Flask-Migrate==2.7.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.14.3
geojson==2.5.0
greenlet==1.0.0
gunicorn==20.0.4
haversine==2.3.0
idna==3.1
itsdangerous==1.1.0
Jinja2==2.11.3
kafka==1.3.5
kafka-python==2.0.2
Mako==1.1.4
MarkupSafe==1.1.1
networkx==2.5
numpy==1.20.1
pip==21.1.2
psycopg2==2.8.6
python-dateutil==2.8.1
python-decouple==3.4
python-editor==1.0.4
setuptools==47.1.0
six==1.15.0
SQLAlchemy==1.4.2
Werkzeug==1.0.1
WTForms==2.3.3
mysql-connector==2.2.9
Flask-Caching==1.10.1
Flask-Session==0.3.2
cachelib==0.1.1
install==1.3.4
mysql-client==0.0.1
mysqlclient==1.4.6
